Renamed countObjects to getBuildings and moved it to library
